refactor: similar blocks code
