+++ src/styles/dialog-media.scss
@@ -1,5 +1,5 @@
 .dialog-media-wrapper {
-  background-color: rgba(0, 0, 0, 0.2);
+  background-color: rgba(0, 0, 0, 0.1);
   position: fixed;
   height: 100%;
   width: 100vw;
@@ -14,7 +14,7 @@
     background-color: #fff;
     box-shadow: 0 3px 7px rgba(0, 0, 0, 0.3);
 
-    width: 40%;
+    width: 35%;
 
     &__header {
       padding: 1rem 1rem 0.6rem 1rem;
